requirements.txt
annotated-types==0.7.0
anyio==4.10.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
dash==3.2.0
dash-bootstrap-components==2.0.4
distro==1.9.0
Flask==3.1.2
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.179.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.74.0
grpcio-status==1.71.2
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.30.0
httpx==0.28.1
idna==3.10
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
MarkupSafe==3.0.2
narwhals==2.2.0
nest-asyncio==1.6.0
numpy==2.3.2
openai==1.102.0
packaging==25.0
pandas==2.3.2
plotly==6.3.0
proto-plus==1.26.1
protobuf==5.29.5
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic_core==2.33.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
requests==2.32.5
retrying==1.4.2
rsa==4.9.1
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
Werkzeug==3.1.3
zipp==3.23.0
